+++ CHANGES.md
@@ -1,4 +1,28 @@
-Changes in RiotX 0.19.0 (2020-XX-XX)
+Changes in RiotX 0.20.0 (2020-XX-XX)
+===================================================
+
+Features ✨:
+ -
+
+Improvements 🙌:
+ -
+
+Bugfix 🐛:
+ - Sometimes the same device appears twice in the list of devices of a user (#1329)
+
+Translations 🗣:
+ -
+
+SDK API changes ⚠️:
+ -
+
+Build 🧱:
+ -
+
+Other changes:
+ -
+
+Changes in RiotX 0.19.0 (2020-05-04)
 ===================================================
 
 Features ✨:
@@ -51,7 +75,7 @@ Bugfix 🐛:
  - Fix bad Shield Logic for DM (#963)
 
 Translations 🗣:
- -
+ - Weblate now create PR directly to RiotX GitHub project
 
 SDK API changes ⚠️:
  - Increase targetSdkVersion to 29